Splits the parser test code path into 3 test suites so one can just run one of them or cache test results easier
